refactored generator methods to utils class
